refactor: replace duplicate dataset config model
