+++ aineko/models/config_schema.py
@@ -5,19 +5,14 @@ from typing import Dict, Optional
 
 from pydantic import BaseModel, Field
 
+from aineko.models.dataset_config_schema import DatasetConfig
+
 
 class Config(BaseModel):
     """Config model."""
 
     class Pipeline(BaseModel):
         """Pipeline model."""
-
-        class Dataset(BaseModel):
-            """Dataset model."""
-
-            type: str
-            location: Optional[str] = None
-            params: Optional[dict] = None
 
         class Node(BaseModel):
             """Node model."""
@@ -31,6 +26,6 @@ class Config(BaseModel):
         name: str
         default_node_settings: Optional[dict] = None
         nodes: Dict[str, Node]
-        datasets: Dict[str, Dataset]
+        datasets: Dict[str, DatasetConfig]
 
     pipeline: Pipeline
